refactor(Lotto Purchase Amount): 파일의 위치를 객체지향적으로 옮겨라
바꾼이유:
- lotto winning number 패키지를 만든 이유와 동일
