+++ src/main/java/lottoPurchaseAmount/LottoPurchaseAmountValidator.java
@@ -1,4 +1,4 @@
-package validate;
+package lottoPurchaseAmount;
 
 import convert.SingleStringToNumConverter;
